requirements.txt
# requirements.txt
streamlit>=1.25.0
numpy>=1.24.0
pandas>=1.5.0
Pillow>=9.4.0
onnxruntime>=1.15.0
plotly>=5.15.0
torch>=2.0.0
torchvision>=0.15.0
